# config-control-service/spec/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Config Control Service API",
    "description": "Centralized configuration management and drift detection service.\n\n**Features:**\n- Service instance heartbeat tracking\n- Configuration drift detection and reporting\n- Service discovery integration with Consul\n- Config refresh orchestration via Kafka\n- Real-time monitoring and alerts\n\n**Architecture:**\n- Hexagonal architecture with distinct API, Application, Domain, and Infrastructure layers\n- MongoDB for persistence\n- Redis for caching\n- Kafka for event broadcasting\n",
    "contact": {
      "name": "Platform Team",
      "email": "platform@example.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "development server"
    }
  ],
  "security": [
    {
      "oauth2_auth_code": [],
      "oauth2_password": []
    }
  ],
...
  "paths": {
    "/api/service-instances/{instanceId}": {
      "get": {
        "tags": [
          "Service Instances"
        ],
        "summary": "Get service instance by ID",
        "description": "Retrieve a specific service instance by instance ID.\n\n**Access Control:**\n- Team members: Can view instances of services owned by their team\n- Shared access: Can view instances of services shared with their team\n- SYS_ADMIN: Can view all instances\n",
        "operationId": "findByIdServiceInstance",
        "parameters": [
          {
            "name": "instanceId",
            "in": "path",
            "description": "Instance ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "payment-dev-1"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Service instance found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceInstanceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Service instance not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2_auth_code": []
          },
          {
            "oauth2_password": []
          }
        ]
      },
      "put": {
        "tags": [
          "Service Instances"
        ],
        "summary": "Update service instance",
        "description": "Update an existing service instance.\n\n**Required Permissions:**\n- Team members: Can update instances of services owned by their team\n- SYS_ADMIN: Can update any instance\n- Updates include configuration hash, status, and drift information\n",
        "operationId": "updateServiceInstance",
        "parameters": [
          {
            "name": "instanceId",
            "in": "path",
            "description": "Instance ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "payment-dev-1"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceInstanceUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Service instance not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Service instance updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceInstanceResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2_auth_code": []
          },
          {
            "oauth2_password": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Service Instances"
        ],
        "summary": "Delete service instance",
        "description": "Delete a service instance permanently.\n\n**Required Permissions:**\n- Team members: Can delete instances of services owned by their team\n- SYS_ADMIN: Can delete any instance\n- This action is irreversible\n",
        "operationId": "deleteServiceInstance",
        "parameters": [
          {
            "name": "instanceId",
            "in": "path",
            "description": "Instance ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "payment-dev-1"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Service instance deleted successfully"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Service instance not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2_auth_code": []
          },
          {
            "oauth2_password": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ServiceInstanceUpdateRequest": {
        "type": "object",
        "properties": {
          "host": {
            "type": "string",
            "description": "Instance host address",
            "example": "payment-dev-1.internal.company.com"
          },
          "port": {
            "type": "integer",
            "description": "Instance port number",
            "format": "int32",
            "example": 8080
          },
          "environment": {
            "type": "string",
            "description": "Deployment environment",
            "example": "dev",
            "enum": [
              "dev",
              "staging",
              "prod"
            ]
          },
          "version": {
            "type": "string",
            "description": "Service version",
            "example": "1.2.1"
          },
          "configHash": {
            "type": "string",
            "description": "Current configuration hash",
            "example": "def456ghi789"
          },
          "lastAppliedHash": {
            "type": "string",
            "description": "Last applied configuration hash",
            "example": "abc123def456"
          },
          "expectedHash": {
            "type": "string",
            "description": "Expected configuration hash",
            "example": "def456ghi789"
          },
          "hasDrift": {
            "type": "boolean",
            "description": "Whether drift is detected",
            "example": true
          },
          "status": {
            "type": "string",
            "description": "Instance status",
            "example": "HEALTHY",
            "enum": [
              "HEALTHY",
              "UNHEALTHY",
              "DRIFT",
              "UNKNOWN",
              "STALE",
              "HEALTHY",
              "UNHEALTHY",
              "DRIFT",
              "UNKNOWN"
            ]
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "description": "Instance metadata",
              "example": "{\"region\":\"us-east-1\",\"zone\":\"us-east-1a\",\"updated\":\"2024-01-15\"}"
            },
            "description": "Instance metadata",
            "example": {
              "region": "us-east-1",
              "zone": "us-east-1a",
              "updated": "2024-01-15"
            }
          }
        },
        "description": "Service instance update request"
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "A URI reference that identifies the problem type",
            "example": "https://api.example.com/problems/validation-error"
          },
          "title": {
            "type": "string",
            "description": "A short, human-readable summary of the problem type",
            "example": "Validation Failed"
          },
          "status": {
            "type": "integer",
            "description": "The HTTP status code",
            "format": "int32",
            "example": 400
          },
          "detail": {
            "type": "string",
            "description": "A human-readable explanation specific to this occurrence of the problem",
            "example": "Request validation failed"
          },
          "instance": {
            "type": "string",
            "description": "A URI reference that identifies the specific occurrence of the problem",
            "example": "/api/application-services"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp when the error occurred",
            "format": "date-time",
            "example": "2024-01-15T10:30:45.123Z"
          },
          "traceId": {
            "type": "string",
            "description": "Trace ID for correlation across services",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "context": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "description": "Additional context about the error"
            },
            "description": "Additional context about the error"
          },
          "validationErrors": {
            "type": "array",
            "description": "Validation errors (for validation failures)",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        },
        "description": "Standardized error response following RFC-7807 Problem Details for HTTP APIs"
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "description": "The field that failed validation",
            "example": "serviceId"
          },
          "rejectedValue": {
            "type": "object",
            "description": "The value that was rejected",
            "example": "invalid-service-id"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "Service ID must not exceed 100 characters"
          },
          "code": {
            "type": "string",
            "description": "Validation constraint code",
            "example": "Size"
          }
        },
        "description": "Individual validation error details"
      },
      "ServiceInstanceResponse": {
        "type": "object",
        "properties": {
          "serviceName": {
            "type": "string",
            "description": "Name of the service",
            "example": "payment-service"
          },
          "instanceId": {
            "type": "string",
            "description": "Unique instance identifier",
            "example": "payment-dev-1"
          },
          "host": {
            "type": "string",
            "description": "Instance host address",
            "example": "payment-dev-1.internal.company.com"
          },
          "port": {
            "type": "integer",
            "description": "Instance port number",
            "format": "int32",
            "example": 8080
          },
          "environment": {
            "type": "string",
            "description": "Deployment environment",
            "example": "dev",
            "enum": [
              "dev",
              "staging",
              "prod"
            ]
          },
          "version": {
            "type": "string",
            "description": "Service version",
            "example": "1.2.0"
          },
          "configHash": {
            "type": "string",
            "description": "Current configuration hash",
            "example": "abc123def456"
          },
          "lastAppliedHash": {
            "type": "string",
            "description": "Last applied configuration hash",
            "example": "abc123def456"
          },
          "expectedHash": {
            "type": "string",
            "description": "Expected configuration hash",
            "example": "abc123def456"
          },
          "status": {
            "type": "string",
            "description": "Instance status",
            "example": "HEALTHY",
            "enum": [
              "HEALTHY",
              "UNHEALTHY",
              "DRIFT",
              "UNKNOWN",
              "STALE",
              "HEALTHY",
              "UNHEALTHY",
              "DRIFT",
              "UNKNOWN"
            ]
          },
          "lastSeenAt": {
            "type": "string",
            "description": "Last time instance was seen",
            "format": "date-time",
            "example": "2024-01-15T14:30:45.123Z"
          },
          "createdAt": {
            "type": "string",
            "description": "Instance creation timestamp",
            "format": "date-time",
            "example": "2024-01-15T10:30:45.123Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "Instance last update timestamp",
            "format": "date-time",
            "example": "2024-01-15T14:30:45.123Z"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "description": "Instance metadata",
              "example": "{\"region\":\"us-east-1\",\"zone\":\"us-east-1a\"}"
            },
            "description": "Instance metadata",
            "example": {
              "region": "us-east-1",
              "zone": "us-east-1a"
            }
          },
          "hasDrift": {
            "type": "boolean",
            "description": "Whether drift is detected",
            "example": false
          },
          "driftDetectedAt": {
            "type": "string",
            "description": "Timestamp when drift was first detected",
            "format": "date-time",
            "example": "2024-01-15T14:30:45.123Z"
          }
        },
        "description": "Service instance details response"
      },
...
    },
    "securitySchemes": {
      "oauth2_auth_code": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "http://keycloak:8080/realms/config-control/protocol/openid-connect/auth",
            "tokenUrl": "http://keycloak:8080/realms/config-control/protocol/openid-connect/token",
            "scopes": {
              "openid": "OpenID",
              "profile": "Profile",
              "email": "Email"
            }
          }
        }
      },
      "oauth2_password": {
        "type": "oauth2",
        "flows": {
          "password": {
            "tokenUrl": "http://keycloak:8080/realms/config-control/protocol/openid-connect/token",
            "scopes": {
              "openid": "OpenID",
              "profile": "Profile",
              "email": "Email"
            }
          }
        }
      }
    }
  }
}